Handle child exit with background procs
Make sure the reaper exits when its direct child exits even if their are
other processes still running in the container.  The bug is that the pid
returned from wait4 will be 0 when there are still other processes
inside the container so we cannot exit only on ECHILD but also have to
check if our child exited if we receive a 0 pid as well.

Signed-off-by: Michael Crosby <crosbymichael@gmail.com>